DjangoRegister/requirements.txt
Django==2.2
djangorestframework==3.11.0
requests==2.23.0